wrap some long lines, add some more checks to validate

+++ libraries/plugins/custom_operations/include/graphene/custom_operations/custom_operations.hpp
@@ -58,7 +58,7 @@ struct create_htlc_order_operation : chain::base_operation
       optional<string> blockchain_account;
       optional<asset> bitshares_amount;
       optional<string> blockchain_asset;
-      optional<uint32_t> blockchain_amount;
+      optional<uint64_t> blockchain_amount;
       optional<fc::time_point_sec> expiration;
       optional<string> tag;
    };
